% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_catalog.R
\name{catalog_entity_path}
\alias{catalog_entity_path}
\title{Compose entity path in data catalog}
\usage{
catalog_entity_path(
  entity,
  layer = character(),
  bucket = character(),
  version = character(),
  fn_valid = valid_data_catalog_entities
)
}
\arguments{
\item{fn_valid}{}
}
\value{

}
\description{
Compose entity path in data catalog
}
