% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_catalog.R
\name{catalog_get_path}
\alias{catalog_get_path}
\title{Get source path from data catalog}
\usage{
catalog_get_path(
  layer,
  version,
  constructor,
  path = character(),
  con_data_lake = env_data_lake()
)
}
\arguments{
\item{constructor}{Class/instance constructor function}

\item{con}{Data connection}

\item{from}{optional source path. In this case this is simply returned}
}
\value{

}
\description{
Get source path from data catalog
}
