% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{catalog_tableau_global_superstore}
\alias{catalog_tableau_global_superstore}
\title{Query layer path in data entity in data catalog}
\usage{
catalog_tableau_global_superstore(layer = character(), version = character())
}
\arguments{
\item{layer}{(\link{character}) Data layer}

\item{version}{(\link{character}) Version of data entity}
}
\value{

}
\description{
Query layer path in data entity in data catalog
}
