% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlang.R
\name{compose_fn_call}
\alias{compose_fn_call}
\title{Compose function call}
\usage{
compose_fn_call(fn, args)
}
\arguments{
\item{fn}{(\link{character}) Function name}

\item{args}{(\link{list}) List of arguments}
}
\value{

}
\description{
Compose function call
}
\examples{
foo <- function(x, y) x + y
call <- compose_fn_call("foo", list(x = 1, y = 1))
eval(call)
}
