% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{con_add_class}
\alias{con_add_class}
\title{Add class information to connection object}
\usage{
con_add_class(con, class_fn = new_con_fs_arrow)
}
\arguments{
\item{class_fn}{}
}
\value{

}
\description{
Add class information to connection object
}
