% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{con_add_path}
\alias{con_add_path}
\title{Add path components to connection object}
\usage{
con_add_path(con, path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Add path components to connection object
}
