% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_con.R
\name{con_fs_csv}
\alias{con_fs_csv}
\title{Constructor}
\usage{
con_fs_csv(x = env_data_lake())
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Constructor
}
\examples{
(x <- con_fs_csv("a/b/c"))
x \%>\% class()
}
