% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_con.R
\name{con_fs_rds}
\alias{con_fs_rds}
\title{Constructor}
\usage{
con_fs_rds(x = env_data_lake())
}
\arguments{
\item{x}{(\link{con_fs}) File system path that inherits from \link{con_fs}}
}
\value{

}
\description{
Constructor
}
\examples{
(x <- con_fs_rds("a/b/c"))
x \%>\% class()
}
