% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read.R
\name{con_read.con_fs_arrow}
\alias{con_read.con_fs_arrow}
\title{Read}
\usage{
\method{con_read}{con_fs_arrow}(
  con = new_con_fs_rds(),
  select = character(),
  where = character(),
  limit = numeric(),
  fn_trans = identity,
  schema = NULL
)
}
\arguments{
\item{schema}{}
}
\value{

}
\description{
Read
}
