% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read.R
\name{con_read.con_fs_rds}
\alias{con_read.con_fs_rds}
\title{Read}
\usage{
\method{con_read}{con_fs_rds}(
  con = new_con_fs_rds(),
  select = character(),
  where = character(),
  limit = numeric(),
  fn_trans = identity
)
}
\arguments{
\item{con}{}
}
\value{

}
\description{
Read
}
