% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_write.R
\name{con_write.con_fs_arrow}
\alias{con_write.con_fs_arrow}
\title{Write data to connection}
\usage{
\method{con_write}{con_fs_arrow}(
  data,
  con,
  layer,
  version,
  constructor,
  to = character(),
  group_by = list(),
  ...
)
}
\arguments{
\item{constructor}{}

\item{to}{(\link{character}) Optional overwrite of destination. Not recommended!}
}
\value{

}
\description{
Write data to connection
}
