% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{construct_class_path}
\alias{construct_class_path}
\title{Construct class path}
\usage{
construct_class_path(x, subclass, prefix = TRUE, sep = "_")
}
\arguments{
\item{x}{\link{ANY}}

\item{subclass}{\link{character} Name of desired subclass. Can be a vector of
length > 1}

\item{prefix}{\link{logical}
\itemize{
\item \code{TRUE} (default): \code{subclass} is prefixed with first element of class path of \code{x}
\item \code{FALSE}: \code{subclass} becomes stand-alone entity within class path
}}
}
\value{

}
\description{
Construct class path
}
\examples{
construct_class_path(x = logical(), "sub")

construct_class_path(x = logical(), c("sub_sub", "sub"))

construct_class_path(
    x = numeric(),
    subclass = "sub",
    prefix = FALSE
)
}
