% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlang.R
\name{curry_fn}
\alias{curry_fn}
\title{Curry function}
\usage{
curry_fn(body)
}
\arguments{
\item{body}{Function body}
}
\value{

}
\description{
Curry function
}
\examples{
foo <- function(x, y) x + y
call <- compose_fn_call("foo", list(x = 1, y = 1))
fn <- call \%>\% curry_fn()
fn()
}
