% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{data}
\alias{data}
\title{Constructor}
\usage{
data(x = tibble::tibble())
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Constructor
}
\examples{
(x <- data(tibble::tibble(x = 1)))
x \%>\% class()
}
