% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{data_tableau_global_superstore}
\alias{data_tableau_global_superstore}
\title{Constructor}
\usage{
data_tableau_global_superstore(
  layer = valid_data_layers("03"),
  version = "v1",
  from = character()
)
}
\arguments{
\item{from}{(\link{character})}

\item{con}{(\link{con_})}
}
\value{

}
\description{
Constructor
}
\examples{
x <- con_fs_arrow() \%>\%
  con_add_path(valid_data_layers("02")) \%>\%
  data_tableau_superstore()
x \%>\% dplyr::glimpse()
x \%>\% class()
}
