% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_data_catalog_version}
\alias{env_data_catalog_version}
\title{Env shorthand}
\usage{
env_data_catalog_version(
  devops_env = Sys.getenv("LAKER_DEVOPS_ENV", pops::get_option("devops_env"))
)
}
\arguments{
\item{devops_env}{}
}
\value{

}
\description{
Env shorthand
}
\examples{
env_data_catalog_version()
}
