% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_con.R
\name{factory_con}
\alias{factory_con}
\title{Factory for con constructor}
\usage{
factory_con(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Factory for con constructor
}
\examples{
fn <- "test.xlsx" \%>\% factory_con()
(con <- fn())
con \%>\% class()
}
