% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{factory_conf_entity}
\alias{factory_conf_entity}
\title{Factory for curried function (\verb{conf_entity_*})}
\usage{
factory_conf_entity(
  data,
  version = "v1",
  layer = "global",
  path_sub = "columns",
  prefix = "^data_"
)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Factory for curried function (\verb{conf_entity_*})
}
