% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{factory_constructor_internal}
\alias{factory_constructor_internal}
\title{Factory for con constructor}
\usage{
factory_constructor_internal(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Factory for con constructor
}
\examples{
fn <- "test.xlsx" \%>\% factory_con()
(con <- fn())
con \%>\% class()
}
