% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{fs_create_symlink}
\alias{fs_create_symlink}
\title{Create symlinc}
\usage{
fs_create_symlink(
  original = "~/data/dev/",
  symlink = "datalake/",
  subdirs = TRUE,
  is_interactive = interactive()
)
}
\arguments{
\item{original}{(\link{character}) Path to original directory to symlink.}

\item{symlink}{(\link{character}) Path to directory to use as symlink.}

\item{subdirs}{(\link{logical}) Link subdirectories of \code{original} into \code{symlink}
(\code{TRUE}) or link \code{original} itself into \code{symlink} (FALSE)?}

\item{is_interactive}{[\link{logical}] Interqctive mode yes/no}
}
\value{

}
\description{
Create symlinc
}
\examples{
\dontrun{
fs_create_symlink(
    original = "~/data/dev/",
    symlink = "datalake/"
)
}
}
