% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{fs_handle_read_path}
\alias{fs_handle_read_path}
\title{Handle file path for read methods}
\usage{
fs_handle_read_path(con = character(), strict = TRUE)
}
\arguments{
\item{con}{(\link{ANY})3 Connection object}

\item{strict}{(\link{logical})3 \link{TRUE}: File must exist | \link{FALSE}: File does not
have to exist}

\item{from}{(\link{character})3 File path}
}
\value{

}
\description{
Testing lists:
\itemize{
\item TRUE: File must exist
\item FALSE: File does not have to exist
}
}
\examples{
try(fs_handle_read_path(con_fs_csv("test.csv")))
fs_handle_read_path(con_fs_xls("test.xls"), strict = FALSE)
fs_handle_read_path(con_fs_xlsx("test.xlsx"), strict = FALSE)
fs_handle_read_path(con_fs_rds("test.Rds"), strict = FALSE)
fs_handle_read_path(con_fs_arrow("test.arrow"), strict = FALSE)
}
