% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laker-package.R
\docType{package}
\name{laker-package}
\alias{laker}
\alias{laker-package}
\title{laker: Local data lake framework}
\description{
A systematic approach to a local data lake framework
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io}

}
\keyword{internal}
