% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_ingest}
\alias{layer_ingest}
\title{Ingest data from raw to tidy layer}
\usage{
layer_ingest(constructor, version = "v1")
}
\arguments{
\item{constructor}{(\link{function}) Constructor function}

\item{version}{(\link{character}) Data entity version}
}
\value{

}
\description{
Ingest data from raw to tidy layer
}
\examples{
layer_ingest(
    constructor = data_tableau_global_superstore
)
}
