% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read.R
\name{layer_read}
\alias{layer_read}
\title{Read from layer}
\usage{
layer_read(constructor, layer = valid_data_layers("02"), version = "v1")
}
\arguments{
\item{version}{}
}
\value{

}
\description{
Convenience wrapper around call to \link{con_read}.
}
