% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{link_data_lake}
\alias{link_data_lake}
\title{Link remote data lake to local directory.}
\usage{
link_data_lake(
  path_data_lake = "~/data/dev/",
  path = "data/",
  subdirs = TRUE,
  is_interactive = interactive()
)
}
\arguments{
\item{path_data_lake}{(\link{character}) Path to remote data lake.}

\item{path}{(\link{character}) Path within the package project to link to
\code{path_data_lake} via symbolic linking}

\item{subdirs}{(\link{logical}) Link subdirectories of \code{original} into \code{symlink}
(\code{TRUE}) or link \code{original} itself into \code{symlink} (FALSE)?}

\item{is_interactive}{[\link{logical}] Interactive mode yes/no}
}
\value{

}
\description{
Note: \code{Remote} is used in the sense of "a directory on the local file system
but outside the package project".
}
\examples{
\dontrun{
link_data_lake(
    path_data_lake = "~/data/dev/",
    path = "datalake/"
)
}
}
