% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_con.R
\name{new_con_fs_arrow}
\alias{new_con_fs_arrow}
\title{Internal constructor}
\usage{
new_con_fs_arrow(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Internal constructor
}
\examples{
(x <- new_con_fs_arrow("a/b/c"))
x \%>\% class()
}
