% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{new_data}
\alias{new_data}
\title{Internal constructor}
\usage{
new_data(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Internal constructor
}
\examples{
(x <- new_data(tibble::tibble(x = 1)))
x \%>\% class()
}
