% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{new_data_tableau_global_superstore}
\alias{new_data_tableau_global_superstore}
\title{Internal constructor}
\usage{
new_data_tableau_global_superstore(x = tibble::tibble())
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Take class path of \link{new_data_tableau} and add prefixed subclass \code{base}.
}
\examples{
(x <- new_data_tableau_superstore(tibble::tibble(x = 1)))
x \%>\% class()
}
