% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{restore_class}
\alias{restore_class}
\title{Restore class information}
\usage{
restore_class(data, to)
}
\arguments{
\item{to}{}
}
\value{
Object with identical class path as \code{to}
}
\description{
Restore class information
}
