% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_data_layers}
\alias{valid_data_layers}
\title{Title}
\usage{
valid_data_layers(
  layer = character(),
  reverse = FALSE,
  flip = FALSE,
  df = FALSE
)
}
\arguments{
\item{df}{(\link{logical}) Return data frame (else vector)?}
}
\value{

}
\description{
Title
}
