% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_dt_arg_selection}
\alias{valid_dt_arg_selection}
\title{Valid DT argument: selection}
\usage{
valid_dt_arg_selection(..., reverse = FALSE, flip = FALSE)
}
\arguments{
\item{flip}{}
}
\value{

}
\description{
See \link[DT:datatable]{DT::datatable}
}
