% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_dt_extensions}
\alias{valid_dt_extensions}
\title{Valid DT extensions}
\usage{
valid_dt_extensions(..., reverse = FALSE, flip = FALSE)
}
\arguments{
\item{flip}{}
}
\value{

}
\description{
See https://datatables.net/extensions/index
}
