% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_con.R
\name{validate_con_fs_csv}
\alias{validate_con_fs_csv}
\title{Validator for internal constructor}
\usage{
validate_con_fs_csv(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Validator for internal constructor
}
