% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_data.R
\name{validate_data_tableau_global_superstore}
\alias{validate_data_tableau_global_superstore}
\title{Validator internal constructor}
\usage{
validate_data_tableau_global_superstore(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Validator internal constructor
}
