% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{fs_create_symlink}
\alias{fs_create_symlink}
\title{Create symlinc}
\usage{
fs_create_symlink(
  original = "~/data/dev/",
  symlink = "data/",
  sub_dirs = FALSE
)
}
\arguments{
\item{sub_dirs}{}
}
\value{

}
\description{
Create symlinc
}
